examples_and_demos/streamlit_requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
sqlite3
asyncio
pathlib
mlflow>=2.8.0
uuid
datetime
json
logging